+++ odoo/local-src/specific_stock/models/__init__.py
@@ -1,7 +1,6 @@
 from . import product_category
 from . import product_template
 from . import res_partner
-from . import stock_backorder_reason
 from . import stock_picking
 from . import stock_production_lot
 from . import stock_location