+++ odoo/local-src/specific_stock/wizards/__init__.py
@@ -1,3 +1,2 @@
-from . import stock_backorder_choice
 from . import stock_receive_lot
 from . import put_in_pack_helper