+++ odoo/local-src/stock_picking_backorder/models/stock_backorder_reason.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 # Copyright 2017 Camptocamp SA
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+# Copyright 2018 Jacques-Etienne Baudoux (BCIM sprl) <je@bcim.be>
+# License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl)
 
 from odoo import fields, models
 
@@ -12,30 +13,26 @@ class StockBackorderReason(models.Model):
     name = fields.Char(
         string='Name',
         required=True,
+        translate=True,
     )
-
     backorder_action_to_do = fields.Selection(
         selection=[
             ('create', 'Create backorder'),
             ('cancel', 'Cancel backorder'),
             (
                 'use_partner_option',
-                'Use partner option (Purchase back order accepted)'
+                'Use partner option (Purchase backorder accepted)'
             ),
         ],
         string='Backorder action to do',
     )
-
+    keep_grn = fields.Boolean(
+        'Keep GRN on backorder')
     is_helpdesk_ticket_to_create = fields.Boolean(
-        string='Is helpdesk ticket to create?',
+        string='Create helpdesk ticket',
     )
-
     helpdesk_ticket_reason_id = fields.Many2one(
         comodel_name='helpdesk.ticket.reason',
         string='Helpdesk ticket reason',
-    )
-
-    helpdesk_ticket_default_name = fields.Char(
-        string='Helpdesk ticket defaut name',
-        help='The default name to use for create a helpdesk ticket.',
+        ondelete='restrict',
     )