+++ odoo/local-src/stock_picking_backorder/tests/test_picking_backorder.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 # Copyright 2017 Camptocamp SA
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+# Copyright 2018 Jacques-Etienne Baudoux (BCIM sprl) <je@bcim.be>
+# License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl)
 
 from odoo.tests.common import TransactionCase, post_install, at_install
 
@@ -52,12 +53,24 @@ class TestPickingBackorder(TransactionCase):
     @post_install(True)
     @at_install(False)
     def test_1_purchase_picking_backorder_create_backorder_no_helpdesk(self):
+        # Define helpdesk ticket values
+        ticket_reason = self.helpdesk_ticket_reason_model.create({
+            'name': 'Unittest helpdesk ticket reason',
+        })
 
         def test():
             # Define the backorder behavior on partner
             self.partner.is_purchase_back_order_accepted = (
                 backorder_accepted
             )
+
+            # Define backorder reason
+            backorder_reason = self.backorder_reason_model.create({
+                'name': 'Unittest backorder',
+                'backorder_action_to_do': backorder_action,
+                'is_helpdesk_ticket_to_create': helpdesk_needed,
+                'helpdesk_ticket_reason_id': ticket_reason.id,
+            })
 
             # Create picking
             picking = self.stock_picking_model.create({
@@ -71,8 +84,8 @@ class TestPickingBackorder(TransactionCase):
                         'product_id': self.product.id,
                         'product_uom_qty': 10,
                         'product_uom': self.product.uom_id.id,
-                        'location_id': self.supplier_location.id,
-                        'location_dest_id': self.stock_location.id,
+                        'location_id': location_id,
+                        'location_dest_id': location_dest_id,
                     })
                 ],
                 'grn_id': self.grn.id,
@@ -88,8 +101,20 @@ class TestPickingBackorder(TransactionCase):
             })
             result = picking.do_new_transfer()
 
-            # Check that the transfer action return no wizard
-            self.assertEqual(result, {})
+            # Check that the transfer action return the good wizard
+            self.assertEqual(
+                result['res_model'],
+                'stock.backorder.choice'
+            )
+
+            # Create backorder choice wizard and execute it
+            wizard = self.backorder_choice_model.with_context(
+                result['context']
+            ).create({
+                'reason_id': backorder_reason.id,
+                'helpdesk_ticket_description': 'test',
+            })
+            wizard.apply()
 
             # Search created backorder
             backorder = self.stock_picking_model.search([
@@ -106,17 +131,40 @@ class TestPickingBackorder(TransactionCase):
             # Check backorder values
             self.assertEqual(len(backorder), 1)
             self.assertEqual(backorder.move_lines.product_uom_qty, 7)
+            keep_backorder = (
+                backorder_action == 'create' or
+                (
+                    backorder_action == 'use_partner_option' and
+                    backorder_accepted
+                )
+            )
             self.assertEqual(
                 backorder.state,
-                'confirmed' if backorder_accepted else 'cancel'
+                'assigned' if keep_backorder else 'cancel'
+            )
+
+            # Check helpdesk ticket creation
+            ticket = self.helpdesk_ticket_model.search([
+                ('stock_picking_id', '=', picking.id),
+            ])
+            if helpdesk_needed:
+                self.assertEqual(len(ticket), 1)
+                self.assertEqual(ticket.partner_id, self.partner)
+                self.assertEqual(
+                    ticket.helpdesk_ticket_reason_id,
+                    ticket_reason
                 )
+            else:
+                self.assertEqual(len(ticket), 0)
 
         # Test all cases
         picking_type_id = self.ref('stock.picking_type_in')
-        location_id = self.ref('stock.stock_location_customers')
+        location_id = self.supplier_location.id
         location_dest_id = self.stock_location.id
-        for backorder_accepted in [False, True]:
-            test()
+        for backorder_action in ['create', 'cancel', 'use_partner_option']:
+            for backorder_accepted in [False, True]:
+                for helpdesk_needed in [False, True]:
+                    test()
 
     @post_install(True)
     @at_install(False)